requirements.txt
Flask
google-cloud-documentai
google-cloud-storage
Pillow
Flask-WTF
gunicorn
Werkzeug
python-dotenv